Extract Storage class for index management
Separate the index document operations from the index setup and removal.
