+++ lib/elasticsearch/extensions/documents/indexer.rb
@@ -2,30 +2,11 @@ module Elasticsearch
   module Extensions
     module Documents
       class Indexer
-        attr_reader :client, :config
+        attr_reader :client, :storage
 
         def initialize(options = {})
           @client = options.fetch(:client) { Documents.client }
-          @config = options.fetch(:config) { Documents.configuration }
-        end
-
-        def drop_index
-          client.indices.delete(index: config.index_name) if index_exists?
-        end
-
-        def index_exists?
-          client.indices.exists(index: config.index_name)
-        end
-
-        def create_index
-          client.indices.create(index: config.index_name, body: index_definition) unless index_exists?
-        end
-
-        def index_definition
-          {}.tap do |body|
-            body[:settings] = config.settings if config.settings
-            body[:mappings] = config.mappings if config.mappings
-          end
+          @storage = options.fetch(:storage) { Storage.new }
         end
 
         def index(payload)
@@ -47,8 +28,8 @@ module Elasticsearch
         def reindex(options = {}, &block)
           force_create = options.fetch(:force_create) { false }
 
-          drop_index if force_create
-          create_index
+          storage.drop_index if force_create
+          storage.create_index
 
           block.call(self) if block_given?
         end
